+++ RELEASE_NOTES.md
@@ -17,11 +17,16 @@ This release prepares the library for the upcoming major version bump.
 
 This release prepares the various input set utilities for the future removal of the input set provider concept and for converting `InputSet` to a struct. 
 
-`v8.0` will be added to the docs, everywhere structural or architectural changes will be made. THe aim is to clean up and simplify the library, which today has too many types and concepts. 
+`v8.0` will be added to the docs, everywhere structural or architectural changes will be made. The aim is to clean up and simplify the library, which today has too many types and concepts. 
 
 ### ✨ New Features
 
 * `KeyboardReturnKeyType` has new `continue`, `emergencyCall` and `route` types.
+* `StandardKeyboardLayoutProvider` has a new default initializer.
+
+### 💡 Adjustments
+
+* `StandardKeyboardLayoutProvider` no longer requires an input set provider.
 
 ### 🗑️ Deprecations